review: clippy clean both workspaces, ABI-qualified packed repr, panic=abort dev
- clippy --all-targets: 0 warnings in root + sdk/rust (fixed Error::other,
  &PathBuf->&Path signatures, too_many_arguments/foo in tests, dead fields)
- generated Rust bindings: #[repr(packed)] -> #[repr(C, packed)] so recovered
  layouts are ABI-qualified; regenerated + stability tests updated
- sdk/rust dev profile: panic=abort so the host build of the no_std device
  staticlib compiles; device panic handler gated on cfg(not(test)) to avoid a
  duplicate panic_impl under cargo test
- extern-C module entry points: #[allow(clippy::not_unsafe_ptr_arg_deref)]
  documented (the loader calls them as plain fn pointers)

full verification: root 61 tests, sdk/rust 55 tests, host C tests, cross-thumb,
C+Rust module cross-build verifier PASS; scripts/ci.sh all 6 gates green

+++ tools/elf-verifier/src/verifier.rs
@@ -152,7 +152,7 @@ impl<'a> Verifier<'a> {
             for (_offset, reloc) in sec.relocations() {
                 report.summary.relocation_count += 1;
                 let rt: u32 = match reloc.flags() {
-                    RelocationFlags::Elf { r_type } => r_type as u32,
+                    RelocationFlags::Elf { r_type } => r_type,
                     _ => 0,
                 };
                 // Only enforce if the loader profile lists relocations.
